removing more old stuff
